Factored out idb functions, loading json into db on init #92
