+++ src/elm/Data/Stim.elm
@@ -6,3 +6,8 @@ import Types exposing (..)
 defaultStim : Stim
 defaultStim =
     Stim "" Chest "" "" Nothing False ""
+
+
+defaultDBStim : DBStim
+defaultDBStim =
+    DBStim "Max" "Chest" "Max2" "Max3" "" False "Max4"
