+++ src/elm/Types.elm
@@ -84,6 +84,17 @@ type alias Stim =
     }
 
 
+type alias DBStim =
+    { stimId : String
+    , bodyPart : String
+    , stimName : String
+    , instructions : String
+    , videoUrl : String
+    , shared : Bool
+    , userId : String
+    }
+
+
 type BodyPart
     = Head
     | Face
