+++ src/js/flickity-handlers.js
