+++ src/js/hotspots-handlers.js
