+++ src/js/idb-helpers.js
@@ -1,3 +1,5 @@
+import Dexie from 'dexie';
+
 const createTables = db => {
   return db.version(1).stores({
     user: 'userId, avatar, skinColour, name',
@@ -5,6 +7,12 @@ const createTables = db => {
     logs:
       '++id, stimId, timeTaken, preFace, postFace, preFeelings, postFeelings, dateTime'
   });
+};
+
+const createDB = () => {
+  const db = new Dexie('stimmy_things');
+  createTables(db);
+  return db;
 };
 
 const createOrUpdateUser = (db, { userId, avatar, skinColour, name }) => {
@@ -68,9 +76,10 @@ const getAllTheData = db => {
 };
 
 export default {
-  createTables,
   createOrUpdateUser,
   getUser,
+  createDB,
+  getAllStims,
   addStim,
   addLog,
   getAllLogs,
