+++ src/json/defaultStims.json
@@ -1,0 +1,20 @@
+[
+  {
+    "stimId": "12345",
+    "stimName": "Jaguar breathing",
+    "bodyPart": "Chest",
+    "instructions": "1. then 2. then 3. then 4.",
+    "videoSrc": "./website.js",
+    "userId": "12345678765432",
+    "shared": false
+  },
+  {
+    "stimId": "678910",
+    "stimName": "Foot Stomp",
+    "bodyPart": "Feet",
+    "instructions": "5. then 6 then 7 then 8",
+    "videoSrc": "./stuff.js",
+    "userId": "1356889538475",
+    "shared": false
+  }
+]
